refactor: Avoid ambiguity problems with clippy
